chore: moves paging types to seperate file
