+++ src/actions/images/collections/read.ts
@@ -3,7 +3,8 @@ import { createActionError } from '@/actions/error'
 import { readImageCollection, readImageCollectionsPage, readSpecialImageCollection } from '@/server/images/collections/read'
 import { safeServerCall } from '@/actions/safeServerCall'
 import { SpecialCollection } from '@prisma/client'
-import type { ActionReturn, ReadPageInput } from '@/actions/Types'
+import type { ActionReturn } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 import type { ImageCollection } from '@prisma/client'
 import type {
     ExpandedImageCollection,
