+++ src/actions/news/read.ts
@@ -2,7 +2,8 @@
 import { safeServerCall } from '@/actions/safeServerCall'
 import { readNews, readNewsCurrent, readOldNewsPage } from '@/server/news/read'
 import type { ExpandedNewsArticle, NewsCursor, SimpleNewsArticle } from '@/server/news/Types'
-import type { ActionReturn, ReadPageInput } from '@/actions/Types'
+import type { ActionReturn } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 
 export async function readOldNewsPageAction<const PageSize extends number>(
     readPageImput: ReadPageInput<PageSize, NewsCursor>
