+++ src/actions/omegaquotes/read.ts
@@ -3,7 +3,8 @@ import { safeServerCall } from '@/actions/safeServerCall'
 import { createActionError } from '@/actions/error'
 import { getUser } from '@/auth/getUser'
 import { readQuotesPage } from '@/server/omegaquotes/read'
-import type { ActionReturn, ReadPageInput } from '@/actions/Types'
+import type { ActionReturn } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 import type { OmegaquoteCursor, OmegaquoteFiltered } from '@/server/omegaquotes/Types'
 
 export async function readQuotesPageAction<const PageSize extends number>(
