+++ src/actions/users/read.ts
@@ -2,7 +2,8 @@
 import { safeServerCall } from '@/actions/safeServerCall'
 import { readUserPage } from '@/server/users/read'
 import type { UserFiltered, UserDetails, UserCursor } from '@/server/users/Types'
-import type { ActionReturn, ReadPageInput } from '@/actions/Types'
+import type { ActionReturn } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 
 /**
  * A action to read a page of users with the given details (filtering)
