+++ src/context/paging/ImageCollectionPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readImageCollectionsPageAction } from '@/actions/images/collections/read'
-import type { ReadPageInput } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 import type { ImageCollectionCursor, ImageCollectionPageReturn } from '@/server/images/collections/Types'
 
 export type PageSizeImageCollection = 12
