+++ src/context/paging/PagingGenerator.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import React, { createContext, useState, useRef, useEffect } from 'react'
-import type { ActionReturn, Page, ReadPageInput } from '@/actions/Types'
+import type { ActionReturn } from '@/actions/Types'
+import type { ReadPageInput, Page } from '@/server/paging/Types'
 import type { Context as ReactContextType } from 'react'
 
 export type StateTypes<Data, Cursor, PageSize extends number> = {
