+++ src/server/omegaquotes/read.ts
@@ -3,7 +3,7 @@ import { omegaQuoteFilterSelection } from './CofigVars'
 import { cursorPageingSelection } from '@/server/paging/cursorPageingSelection'
 import { prismaCall } from '@/server/prismaCall'
 import prisma from '@/prisma'
-import type { ReadPageInput } from '@/actions/Types'
+import type { ReadPageInput } from '@/server/paging/Types'
 import type { OmegaquoteCursor, OmegaquoteFiltered } from '@/server/omegaquotes/Types'
 
 export async function readQuotesPage<const PageSize extends number>(
