+++ src/server/paging/cursorPageingSelection.ts
@@ -1,5 +1,5 @@
 import 'server-only'
-import type { Page } from '@/actions/Types'
+import type { Page } from '@/server/paging/Types'
 
 /**
  * A function to generate the cursor-paging selection for a given page
